refactored symbols.py
